requirements/requirements.txt
streamlit==1.28.0
pandas==2.1.1
numpy==1.27.4
pydeck==0.8.0
altair==5.0.1
requests==2.31.0
python-dotenv==1.1.1
beautifulsoup4==4.12.2
lxml==4.9.3
google-generativeai==0.1.6
